Thêm kiểm tra những danh mục mới được thêm vào đã tồn tại hay chưa

+++ controllers/category.c.js
@@ -69,35 +69,51 @@ module.exports = {
             let catName = urlObj.searchParams.get("catName");
             let add = urlObj.searchParams.get("add");
             
+
             if (deleteID) {
                 await Category.deleteByID(parseInt(deleteID));
             }
-            if (editID) {
-                await Category.updateCategory(parseInt(editID), catName);
+            if (catName && await Category.checkCatNameExist(catName.toLowerCase().replace(/(?:^|\s)\S/g, function(a) { return a.toUpperCase(); }))){
+                    throw(`Tên danh mục ${catName.toLowerCase().replace(/(?:^|\s)\S/g, function(a) { return a.toUpperCase(); })} đã tồn tại!`);
+            }else {
+                if (editID) {
+                
+                    await Category.updateCategory(parseInt(editID), catName);
+                }
             }
-            if (add) {
-                await Category.addCategory(add);
+            if (add && await Category.checkCatNameExist(add.toLowerCase().replace(/(?:^|\s)\S/g, function(a) { return a.toUpperCase(); }))){
+                throw(`Tên danh mục ${add.toLowerCase().replace(/(?:^|\s)\S/g, function(a) { return a.toUpperCase(); })} đã tồn tại!`);
+            } else {
+                if (add) {
+                    await Category.addCategory(add);
+                }
             }
             let addCatID = urlObj.searchParams.get("addtoID");
             let addItemID = urlObj.searchParams.get("itemID");
             let addItemName = urlObj.searchParams.get("itemName");
-
-            let deleteItemID = urlObj.searchParams.get("deleteItem");
-
             let editItem =  urlObj.searchParams.get("editItem");
             let itemName =  urlObj.searchParams.get("itemName");
             let itemCategory =  urlObj.searchParams.get("catID");
+            let deleteItemID = urlObj.searchParams.get("deleteItem");
 
-            if (editItem && itemName && itemCategory){
-                await Category.updateItem(editItem, itemName, parseInt(itemCategory));
+            if (itemName && await Category.checkItemNameExist(itemName.toLowerCase().replace(/(?:^|\s)\S/g, function(a) { return a.toUpperCase(); })) || addItemName&&await Category.checkItemNameExist(addItemName.toLowerCase().replace(/(?:^|\s)\S/g, function(a) { return a.toUpperCase(); }))){
+                throw(`Tên danh mục ${itemName.toLowerCase().replace(/(?:^|\s)\S/g, function(a) { return a.toUpperCase(); })} đã tồn tại!`);
             }
-            
-            if (addCatID && addItemID && addItemName){
-                await Category.addItem(addItemID, addItemName, parseInt(addCatID));
+            else if (await Category.checkItemIDExist(addItemID)){
+                throw(`Danh mục có ID này đã tồn tại!`);
             }
-            if (deleteItemID) {
-                await Category.deleteItemByID(deleteItemID);
-            }
+            else {
+                if (editItem && itemName && itemCategory){
+                    await Category.updateItem(editItem, itemName, parseInt(itemCategory));
+                }
+                
+                if (addCatID && addItemID && addItemName){
+                    await Category.addItem(addItemID, addItemName, parseInt(addCatID));
+                }
+                if (deleteItemID) {
+                    await Category.deleteItemByID(deleteItemID);
+                }
+            }         
             
             let categories = await Category.allCategory();       
             let categoryItems = await Category.allCategoryItem();
@@ -114,8 +130,8 @@ module.exports = {
                 const items = categoryItems.filter((item) => item.catID === categories.catID);
                 return { ...categories, items };
             });
-            res.render("admin/category/viewCategory", {  layout: 'admin', categories: dataForHbs, error:'Không thể xoá do tồn tại sản phẩm sản phẩm liên quan tới danh mục muốn hoá hiện tại!', title: "Edit" });
-        }
+            res.render("admin/category/viewCategory", {  layout: 'admin', categories: dataForHbs, error, title: "Edit" });
+        }//:'Không thể xoá do tồn tại sản phẩm sản phẩm liên quan tới danh mục muốn hoá hiện tại!'
     },
     renderCatForHome: async (req, res, next) => {
         try{
@@ -131,4 +147,5 @@ module.exports = {
             next(error);
         }
     },
+    
 }
